requirements.txt
langchain
langchain-community
langchain-core
groq
python-dotenv
wikipedia
requests
PyMuPDF
python-docx
